Rename dwr to wrd
