+++ userCode/__init__.py
@@ -19,19 +19,19 @@ import dagster_slack
 import userCode.xlsx.dag as xlsx
 from userCode.util import get_env, slack_error_fn
 import userCode.awqms.dag as awqms
-import userCode.odwr.dag as odwr
+import userCode.wrd.dag as wrd
 import userCode.groundwater.dag as groundwater
 
-assets = load_assets_from_modules([awqms, odwr, groundwater, xlsx])
-asset_checks = load_asset_checks_from_modules([awqms, odwr, groundwater, xlsx])
+assets = load_assets_from_modules([awqms, wrd, groundwater, xlsx])
+asset_checks = load_asset_checks_from_modules([awqms, wrd, groundwater, xlsx])
 
 definitions = Definitions(
     assets=assets,
     asset_checks=asset_checks,
-    jobs=[awqms.awqms_job, odwr.odwr_job, groundwater.groundwater_job, xlsx.xlsx_job],
+    jobs=[awqms.awqms_job, wrd.wrd_job, groundwater.groundwater_job, xlsx.xlsx_job],
     schedules=[
         awqms.awqms_schedule,
-        odwr.odwr_schedule,
+        wrd.wrd_schedule,
         groundwater.groundwater_schedule,
         xlsx.xlsx_schedule,
     ],
