+++ userCode/wrd/dag_test.py
@@ -16,14 +16,14 @@ import requests
 from userCode import definitions
 from userCode.ontology import ONTOLOGY_MAPPING
 from userCode.env import API_BACKEND_URL
-from userCode.odwr.dag import all_metadata, post_station, sta_station
+from userCode.wrd.dag import all_metadata, post_station, sta_station
 from userCode.helper_classes import get_datastream_time_range, MockValues
-from userCode.odwr.lib import (
+from userCode.wrd.lib import (
     assert_no_observations_with_same_iotid_in_first_page,
     generate_oregon_tsv_url,
     parse_oregon_tsv,
 )
-from userCode.odwr.types import StationData
+from userCode.wrd.types import StationData
 from userCode.util import PACIFIC_TIME, from_oregon_datetime, to_oregon_datetime
 
 
@@ -121,7 +121,7 @@ def test_full_pipeline(metadata: list[StationData]):
     test_iow_hash_is_deterministic()
     assert_observations_and_datastreams_empty()
 
-    harvest_job = definitions.get_job_def("harvest_owdp")
+    harvest_job = definitions.get_job_def("harvest_wrd")
 
     instance = DagsterInstance.ephemeral()
